+++ offline-lyrics/src/main/java/dev/olog/offlinelyrics/model/LyricsModel.kt
@@ -1,0 +1,6 @@
+package dev.olog.offlinelyrics.model
+
+data class LyricsModel(
+    val id: Long,
+    val lyrics: String
+)