+++ service-floating/src/main/java/dev/olog/service/floating/OfflineLyricsContentPresenter.kt
@@ -1,18 +1,18 @@
 package dev.olog.service.floating
 
-import dev.olog.core.prefs.AppPreferencesGateway
+import dev.olog.core.gateway.OfflineLyricsGateway
 import dev.olog.offlinelyrics.BaseOfflineLyricsPresenter
 import dev.olog.offlinelyrics.domain.InsertOfflineLyricsUseCase
 import dev.olog.offlinelyrics.domain.ObserveOfflineLyricsUseCase
 import javax.inject.Inject
 
 class OfflineLyricsContentPresenter @Inject constructor(
-    appPreferencesUseCase: AppPreferencesGateway,
+    lyricsGateway: OfflineLyricsGateway,
     observeUseCase: ObserveOfflineLyricsUseCase,
     insertUseCase: InsertOfflineLyricsUseCase
 
 ) : BaseOfflineLyricsPresenter(
-    appPreferencesUseCase,
+    lyricsGateway,
     observeUseCase,
     insertUseCase
 )